cli: move choose_commit from next/prev commands to cli_util
